requirements.txt
jupyterlab
pandas
pygmsh
scipy
pyvista
adios4dolfinx
imageio>=2.31,<3
